Consolidated traversal heading enums into new .builders.heading.py
(cherry picked from commit c3c27f2b4c63802a0bcc6b09a997a48e0fee16cf)
