+++ mupt/interfaces/rdkit/importers.py
@@ -26,7 +26,7 @@ from ...chemistry.smiles import DEFAULT_SMILES_WRITE_PARAMS, SmilesWriteParams
 from ...chemistry.conversion import rdkit_atom_to_element
 
 from ...mupr.primitives import Primitive, PrimitiveHandle
-from ...mupr.connection import TraversalDirection
+from ...builders.heading import TraversalDirection
 
 
 def primitive_from_rdkit_atom(
